pdqr: Work with Custom Distribution Functions

Create, transform, and summarize custom random variables with distribution functions (analogues of 'p*()', 'd*()', 'q*()', and 'r*()' functions from base R). Two types of distributions are supported: "discrete" (random variable has finite number of output values) and "continuous" (infinite number of values in the form of continuous random variable). Functions for distribution transformations and summaries are available. Implemented approaches often emphasize approximate and numerical solutions: all distributions assume finite support and finite values of density function; some methods implemented with simulation techniques.

Details

Excerpt of important documentation:

README and vignettes provide overview of package functionality.

Documentation of meta_*() functions describes implementation details of pdqr-functions.

Documentation of print() and plot() methods describes how you can interactively explore properties of pdqr-functions.

Documentation of new_*() functions describes the process of creating pdqr-functions.

Documentation of as_*() functions describes the process of updating class of pdqr-functions.

Documentation of form_*() functions describes how different transformation functions work. Important pages are for form_trans() and Pdqr methods for S3 group generic functions.

Documentation of summ_*() functions describes how different summary functions work. A good place to start is summ_center().

Documentation of region_*() functions describes functionality to work with regions: data frames defining subset of one dimensional real line.

This package has the following options (should be set by options()):

"pdqr.approx_discrete_n_grid". This single number (default to 1000) determines degree of granularity of how continuous pdqr-function is approximated with discrete one during some complicated tasks. Approximation is done by first using form_regrid() with n_grid argument equal to this option and method = "x", and then form_retype() is used with type = "discrete" and method = "piecelin". Value of this option should be big enough for high accuracy and small enough for high computation speed, for which value 1000 showed to be fairly appropriate.

"pdqr.assert_args". This boolean option (default to TRUE) may be used to turn off sanity checks of function arguments (set it to FALSE), which will somewhat increase general execution speed. Use this option at your own risk in case you are confident that input arguments have correct type and structure.

"pdqr.group_gen.args_new", "pdqr.group_gen.n_sample", "pdqr.group_gen.repair_supp_method". They may be used to customize behavior of methods for S3 group generic functions. See their help page for more information.

See also

Author

Maintainer: Evgeni Chasnovski evgeni.chasnovski@gmail.com (ORCID)