pdqr: Work with Custom Distribution Functions

Create, transform, and summarize custom random variables with distribution functions (analogues of 'p*()', 'd*()', 'q*()', and 'r*()' functions from base R). Two types of distributions are supported: "discrete" (random variable has finite number of output values) and "continuous" (infinite number of values in the form of continuous random variable). Functions for distribution transformations and summaries are available. Implemented approaches often emphasize approximate and numerical solutions: all distributions assume finite support and finite values of density function; some methods implemented with simulation techniques.

Details

Excerpt of important documentation:

README and vignettes provide overview of package functionality.

Documentation of meta_*() functions describes implementation details of pdqr-functions.

Documentation of print() and plot() methods describes how you can interactively explore properties of pdqr-functions.

Documentation of new_*() functions describes the process of creating pdqr-functions.

Documentation of as_*() functions describes the process of updating class of pdqr-functions.

Documentation of form_*() functions describes how different transformation functions work. Important pages are for form_trans() and Pdqr methods for S3 group generic functions.

Documentation of summ_*() functions describes how different summary functions work. A good place to start is summ_center().

Documentation of region_*() functions describes functionality to work with regions: data frames defining subset of one dimensional real line.

See also